This project is a biginner-level data analysis.
This project is a beginner-level data analysis task built using Python and Jupyter Notebook. The main goal of this notebook is to load and explore an anime.csv dataset, which contains information related to various anime titles, such as their names, scores, popularity, and other metadata. This type of analysis is helpful for practicing data handling skills using popular Python libraries like pandas and numpy.